+++ packages/@tinacms/fields/src/Wysiwyg/node-views/index.ts
@@ -20,6 +20,7 @@ import { Node } from 'prosemirror-model'
 import { Decoration, EditorView, NodeView } from 'prosemirror-view'
 
 import { CodeBlockView } from './CodeBlockView'
+import { ImageView } from './ImageView'
 
 export type NodeViews = {
   [name: string]: (
@@ -33,5 +34,8 @@ export type NodeViews = {
 export const nodeViews: NodeViews = {
   code_block(node, view, getPos) {
     return new CodeBlockView(node, view, getPos)
+  },
+  image(node, view, getPos) {
+    return new ImageView(node, view, getPos)
   },
 }