+++ packages/@tinacms/form-builder/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.12-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/form-builder@0.2.11...@tinacms/form-builder@0.2.12-alpha.0) (2020-01-16)
+
+**Note:** Version bump only for package @tinacms/form-builder
+
+
+
+
+
+## [0.2.11](https://github.com/tinacms/tinacms/compare/@tinacms/form-builder@0.2.10...@tinacms/form-builder@0.2.11) (2020-01-14)
+
+**Note:** Version bump only for package @tinacms/form-builder
+
+
+
+
+
+## [0.2.10](https://github.com/tinacms/tinacms/compare/@tinacms/form-builder@0.2.10-alpha.1...@tinacms/form-builder@0.2.10) (2020-01-13)
+
+**Note:** Version bump only for package @tinacms/form-builder
+
+
+
+
+
+## [0.2.10-alpha.1](https://github.com/tinacms/tinacms/compare/@tinacms/form-builder@0.2.10-alpha.0...@tinacms/form-builder@0.2.10-alpha.1) (2020-01-10)
+
+**Note:** Version bump only for package @tinacms/form-builder
+
+
+
+
+
 ## [0.2.9](https://github.com/tinacms/tinacms/compare/@tinacms/form-builder@0.2.9-alpha.1...@tinacms/form-builder@0.2.9) (2019-12-17)
 
 **Note:** Version bump only for package @tinacms/form-builder