+++ packages/@tinacms/teams/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/teams@0.1.2...@tinacms/teams@0.2.0-alpha.0) (2020-01-16)
+
+
+### Features
+
+* **auth:** Add a link to logout from auth screen ([2f090ef](https://github.com/tinacms/tinacms/commit/2f090ef))
+
+
+
+
+
+## [0.1.2](https://github.com/tinacms/tinacms/compare/@tinacms/teams@0.1.2-alpha.0...@tinacms/teams@0.1.2) (2020-01-13)
+
+**Note:** Version bump only for package @tinacms/teams
+
+
+
+
+
 ## [0.1.1](https://github.com/tinacms/tinacms/compare/@tinacms/teams@0.1.1-alpha.1...@tinacms/teams@0.1.1) (2019-12-17)
 
 **Note:** Version bump only for package @tinacms/teams