+++ packages/gatsby-tinacms-json/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.4-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.3...gatsby-tinacms-json@0.8.4-alpha.0) (2020-01-16)
+
+
+### Bug Fixes
+
+* disregard intermittent null vals from gatsby ([d102745](https://github.com/tinacms/tinacms/commit/d102745))
+
+
+
+
+
+## [0.8.3](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.2...gatsby-tinacms-json@0.8.3) (2020-01-14)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+## [0.8.2](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.2-alpha.1...gatsby-tinacms-json@0.8.2) (2020-01-13)
+
+
+### Bug Fixes
+
+* import correct GraphQLString type ([bba2979](https://github.com/tinacms/tinacms/commit/bba2979))
+
+
+
+
+
+## [0.8.2-alpha.1](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.2-alpha.0...gatsby-tinacms-json@0.8.2-alpha.1) (2020-01-10)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
 ## [0.8.1](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.1-alpha.1...gatsby-tinacms-json@0.8.1) (2019-12-17)
 
 **Note:** Version bump only for package gatsby-tinacms-json