+++ packages/react-dismissible/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2-alpha.0](https://github.com/tinacms/tinacms/compare/react-dismissible@1.1.1...react-dismissible@1.1.2-alpha.0) (2020-01-16)
+
+**Note:** Version bump only for package react-dismissible
+
+
+
+
+
+## [1.1.1](https://github.com/tinacms/tinacms/compare/react-dismissible@1.1.0...react-dismissible@1.1.1) (2020-01-14)
+
+**Note:** Version bump only for package react-dismissible
+
+
+
+
+
+# [1.1.0](https://github.com/tinacms/tinacms/compare/react-dismissible@1.1.0-alpha.1...react-dismissible@1.1.0) (2020-01-13)
+
+**Note:** Version bump only for package react-dismissible
+
+
+
+
+
+# [1.1.0-alpha.1](https://github.com/tinacms/tinacms/compare/react-dismissible@1.1.0-alpha.0...react-dismissible@1.1.0-alpha.1) (2020-01-10)
+
+**Note:** Version bump only for package react-dismissible
+
+
+
+
+
 ## [1.0.18](https://github.com/tinacms/tinacms/compare/react-dismissible@1.0.18-alpha.1...react-dismissible@1.0.18) (2019-12-17)
 
 **Note:** Version bump only for package react-dismissible